% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{a}
\alias{a}
\title{：用于描述函数的标题或名称。}
\usage{
：给出函数的用法示例。
}
\value{
：说明函数的返回值。
}
\description{
：提供该函数的简短描述。
}
\details{
：进一步详细说明函数的工作机制和细节。
}
\examples{
：提供函数的使用示例代码。
}
\keyword{：列出与函数相关的关键词，以便在文档中进行索引。}
