% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlier_sample.R
\name{get_outlier_sample}
\alias{get_outlier_sample}
\title{检测表达矩阵中的离群样本}
\usage{
get_outlier_sample(
  expression_data,
  z_threshold = -2,
  plot_hclust = FALSE,
  show_plot = TRUE,
  hclust_filename = NULL,
  connectivity_filename = NULL
)
}
\arguments{
\item{expression_data}{数据框，tpm or fpkm表达矩阵，行为基因，列为样本。}

\item{z_threshold}{数值，用于确定离群样本的 Z 得分阈值，默认值为 -3。}

\item{plot_hclust}{逻辑值，是否绘制样本的层次聚类图，默认 FALSE。}

\item{show_plot}{逻辑值，是否在屏幕上显示绘制的图形，默认 TRUE。}

\item{hclust_filename}{字符串，保存层次聚类图的文件名（包括路径），如果为 NULL 则不保存，默认 NULL。}

\item{connectivity_filename}{字符串，保存样本连接度图的文件名（包括路径），如果为 NULL 则不保存，默认 NULL。}
}
\value{
返回可能的离群样本的名称向量。
}
\description{
该函数用于识别表达矩阵数据中可能的离群样本，并可选择性地绘制和保存相关的图形。
}
\details{
该函数首先计算表达矩阵中样本之间的双相关系数，然后基于此构建样本之间的连接度网络。
通过计算每个样本的连接度的 Z 得分，可以识别出可能的离群样本，即那些 Z 得分低于指定阈值（\code{z_threshold}）的样本。
可以选择绘制并保存层次聚类图和样本连接度图，以便进一步分析。
}
\examples{
\dontrun{
# 示例代码
outliers <- get_outlier_sample(
  expression_data = expr_fpkm,
  z_threshold = -3,
  plot_hclust = TRUE,
  show_plot = TRUE,
  hclust_filename = NULL,
  connectivity_filename = NULL
)
expr_tpm <- expr_tpm[, !colnames(expr_tpm) \%in\% outline_sample]
}
}
\author{
Fan Xingfu
}
\keyword{WGCNA}
\keyword{离群样本检测}
\keyword{表达矩阵}
