% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_median.R
\name{group_by_median}
\alias{group_by_median}
\title{按中位数分组}
\usage{
group_by_median(data, column_name)
}
\arguments{
\item{data}{包含要分组的数值列的数据框}

\item{column_name}{要分组的列名（字符串）}
}
\value{
添加了分组列的新数据框
}
\description{
将指定的数值列按其中位数分为高低两组
}
\details{
此函数计算指定列的中位数，然后创建一个新的列，
将大于中位数的值标记为'high'，小于或等于中位数的值标记为'low'。
新列的名称为原列名加上"_group"后缀。
}
\examples{
# 创建一个示例数据框
data <- data.frame(ID = 1:10, Value = runif(10))
# 按Value列的中位数进行分组
grouped_data <- group_by_median(data, "Value")
print(grouped_data)
}
\author{
Fan Xingfu
}
