% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_run.R
\name{proxy_run}
\alias{proxy_run}
\title{使用代理设置运行函数}
\usage{
proxy_run(
  f,
  http_proxy = "http://localhost:1086",
  https_proxy = "http://localhost:1086"
)
}
\arguments{
\item{f}{要在代理设置下执行的函数。}

\item{http_proxy}{HTTP 代理 URL。默认为 "http://localhost:1086"。}

\item{https_proxy}{HTTPS 代理 URL。默认为 "http://localhost:1086"。}
}
\value{
执行函数的结果，如果发生错误则返回 NULL。
}
\description{
此函数在指定的代理设置下执行给定的函数。
它会临时设置代理环境变量，测试代理，然后运行提供的函数。
}
\examples{
proxy_run(function() {
  httr::GET("https://api.github.com/users/hadley")
})
}
\author{
Fan Xingfu
}
