% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates_func.R
\name{remove_duplicates_func}
\alias{remove_duplicates_func}
\title{去除重复并聚合数据}
\usage{
remove_duplicates_func(data, agg_func = mean)
}
\arguments{
\item{data}{输入的数据框}

\item{agg_func}{聚合函数，对其他列应用的函数，默认为mean}
}
\value{
去重并应用聚合函数后的数据框，第一列作为行名
}
\description{
按第一列分组去除重复行，并对其他列应用指定的聚合函数
}
\details{
此函数首先按第一列对数据进行分组，然后对其他所有列应用指定的聚合函数。
结果中，第一列的唯一值将作为行名，其他列为对应的聚合结果。
}
\examples{
# 创建一个示例数据框
data <- data.frame(ID = c("A", "A", "B", "B"), Value1 = 1:4, Value2 = 5:8)
# 去除重复并计算均值
unique_data_mean <- remove_duplicates_func(data)
print(unique_data_mean)
# 使用最大值
unique_data_max <- remove_duplicates_func(data, max)
print(unique_data_max)
# 使用中位数
unique_data_median <- remove_duplicates_func(data, median)
print(unique_data_median)
# 使用自定义函数（例如，计算范围）
unique_data_range <- remove_duplicates_func(data, function(x) max(x) - min(x))
print(unique_data_range)
}
\author{
Fan Xingfu
}
